@@ -1,52 +1,40 @@
 {
   "name": "sveltekit-template",
-  "version": "0.0.2",
+  "version": "0.0.3",
   "private": true,
   "type": "module",
   "scripts": {
     "dev": "vite dev",
-    "build:site": "vite build",
-    "build:lib": "svelte-kit sync && svelte-package",
+    "build": "vite build",
     "preview": "vite preview",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "test:unit": "vitest",
-    "lint": "eslint . --fix",
-    "format": "prettier --write ./**/*.{js,cjs,ts,html,css,svelte,md,json}"
+    "test": "vitest",
+    "lint": "prettier --check . && eslint .",
+    "format": "prettier --write ."
   },
   "devDependencies": {
-    "@iconify-json/carbon": "^1.1.30",
-    "@sveltejs/adapter-auto": "^3.1.1",
-    "@sveltejs/kit": "^2.5.0",
-    "@sveltejs/package": "^2.2.6",
-    "@sveltejs/vite-plugin-svelte": "^3.0.2",
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.0.1",
-    "@typescript-eslint/parser": "^7.0.1",
-    "@unocss/core": "^0.58.5",
-    "@unocss/eslint-config": "^0.58.5",
-    "@unocss/extractor-svelte": "^0.58.5",
-    "@unocss/preset-attributify": "^0.58.5",
-    "@unocss/preset-icons": "^0.58.5",
-    "@unocss/preset-uno": "^0.58.5",
-    "@unocss/reset": "^0.58.5",
-    "@unocss/transformer-directives": "^0.58.5",
-    "@unocss/transformer-variant-group": "^0.58.5",
-    "@unocss/vite": "^0.58.5",
-    "eslint": "^8.56.0",
+    "@fontsource/fira-mono": "^5.0.13",
+    "@sveltejs/adapter-auto": "^3.2.1",
+    "@sveltejs/kit": "^2.5.10",
+    "@sveltejs/vite-plugin-svelte": "^3.1.1",
+    "@tailwindcss/vite": "4.0.0-alpha.15",
+    "@types/eslint": "^8.56.10",
+    "eslint": "^9.4.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-svelte": "^2.35.1",
-    "eslint-plugin-vitest": "^0.3.22",
-    "jsdom": "^24.0.0",
-    "prettier": "^3.2.5",
-    "prettier-plugin-svelte": "^3.2.1",
-    "svelte": "^4.2.11",
-    "svelte-check": "^3.6.4",
-    "svelte-eslint-parser": "^0.33.1",
-    "tslib": "^2.6.2",
-    "typescript": "^5.3.3",
-    "vite": "^5.1.3",
-    "vitest": "^1.3.0"
+    "eslint-plugin-svelte": "^2.39.0",
+    "globals": "^15.3.0",
+    "prettier": "^3.3.1",
+    "prettier-plugin-svelte": "^3.2.3",
+    "prettier-plugin-tailwindcss": "^0.6.1",
+    "svelte": "5.0.0-next.149",
+    "svelte-check": "^3.8.0",
+    "tailwindcss": "4.0.0-alpha.15",
+    "tslib": "^2.6.3",
+    "typescript": "^5.4.5",
+    "typescript-eslint": "8.0.0-alpha.28",
+    "unplugin-icons": "^0.19.0",
+    "vite": "^5.2.12",
+    "vitest": "^1.6.0"
   }
 }
